+++ src/main/java/de/maxhenkel/easyvillagers/blocks/BreederBlock.java
@@ -58,12 +58,12 @@ public class BreederBlock extends VillagerBlockBase implements ITileEntityProvid
         if (!breeder.hasVillager1() && heldItem.getItem() instanceof VillagerItem) {
             breeder.setVillager1(heldItem.copy());
             ItemUtils.decrItemStack(heldItem, player);
-            TraderBlock.playVillagerSound(worldIn, pos, SoundEvents.ENTITY_VILLAGER_YES);
+            VillagerBlockBase.playVillagerSound(worldIn, pos, SoundEvents.ENTITY_VILLAGER_YES);
             return ActionResultType.SUCCESS;
         } else if (!breeder.hasVillager2() && heldItem.getItem() instanceof VillagerItem) {
             breeder.setVillager2(heldItem.copy());
             ItemUtils.decrItemStack(heldItem, player);
-            TraderBlock.playVillagerSound(worldIn, pos, SoundEvents.ENTITY_VILLAGER_YES);
+            VillagerBlockBase.playVillagerSound(worldIn, pos, SoundEvents.ENTITY_VILLAGER_YES);
             return ActionResultType.SUCCESS;
         } else if (player.isSneaking() && breeder.hasVillager2()) {
             ItemStack stack = breeder.removeVillager2();
@@ -75,7 +75,7 @@ public class BreederBlock extends VillagerBlockBase implements ITileEntityProvid
                     InventoryHelper.spawnItemStack(worldIn, direction.getXOffset() + pos.getX() + 0.5D, pos.getY() + 0.5D, direction.getZOffset() + pos.getZ() + 0.5D, stack);
                 }
             }
-            TraderBlock.playVillagerSound(worldIn, pos, SoundEvents.ENTITY_VILLAGER_CELEBRATE);
+            VillagerBlockBase.playVillagerSound(worldIn, pos, SoundEvents.ENTITY_VILLAGER_CELEBRATE);
             return ActionResultType.SUCCESS;
         } else if (player.isSneaking() && breeder.hasVillager1()) {
             ItemStack stack = breeder.removeVillager1();
@@ -87,7 +87,7 @@ public class BreederBlock extends VillagerBlockBase implements ITileEntityProvid
                     InventoryHelper.spawnItemStack(worldIn, direction.getXOffset() + pos.getX() + 0.5D, pos.getY() + 0.5D, direction.getZOffset() + pos.getZ() + 0.5D, stack);
                 }
             }
-            TraderBlock.playVillagerSound(worldIn, pos, SoundEvents.ENTITY_VILLAGER_CELEBRATE);
+            VillagerBlockBase.playVillagerSound(worldIn, pos, SoundEvents.ENTITY_VILLAGER_CELEBRATE);
             return ActionResultType.SUCCESS;
         } else {
             player.openContainer(new INamedContainerProvider() {